Refactor UnitSkillTabState
